add validation and unit tests to UpdateEntry method

+++ my-diary.Api/Controllers/EntriesController.cs
@@ -58,11 +58,21 @@ namespace my_diary.Api.Controllers
                 return BadRequest($"Entry of Id {id} does not exist");
             } 
           
+            if (!IsValidEntry(entry))
+            {
+                return BadRequest("Invalid Entry");
+            }
+            entry.LastUpdated = DateTimeOffset.UtcNow;
             db.Entries.Remove(oldEntry);
             db.Entries.Add(entry);
 
             return Ok("Successfully Updated");
 
+        }
+
+        private bool IsValidEntry(Entry entry)
+        {
+            return !string.IsNullOrWhiteSpace(entry.Id) || !string.IsNullOrWhiteSpace(entry.Title) || !string.IsNullOrWhiteSpace(entry.Text);
         }
     }
 }
